+++ projects/huskeyfi/backend/.env.example
@@ -1,0 +1,3 @@
+WEBHOOK_SECRET=3WrgwgUqpDkgB3JKfyEwNH2VYmXg
+BTCPAY_URL=BTCPAY_URL
+BRIDGE_CENTER_ID=BTCPAY_STORE_ID
