+++ projects/huskeyfi/backend/src/app.service.ts
@@ -2,10 +2,27 @@ import { Injectable } from '@nestjs/common';
 import { Invoice, InvoiceResponse } from './types';
 import axios from 'axios';
 
+enum WebhookType {
+  InvoiceCreated = 'InvoiceCreated',
+  InvoiceExpired = 'InvoiceExpired',
+  InvoiceReceivedPayment = 'InvoiceReceivedPayment',
+  InvoicePaymentSettled = 'InvoicePaymentSettled',
+  InvoiceProcessing = 'InvoiceProcessing',
+  InvoiceInvalid = 'InvoiceInvalid',
+  InvoiceSettled = 'InvoiceSettled',
+}
+
 @Injectable()
 export class AppService {
   private BTCPAY_URL = process.env.BTCPAY_URL;
   private BRIDGE_CENTER_ID = process.env.BRIDGE_CENTER_ID;
+  private API_KEY = process.env.API_KEY;
+
+  private getAuthHeaders() {
+    return {
+      Authorization: `token ${this.API_KEY}`,
+    };
+  }
 
   getHello(): string {
     return 'Hello Hashkey!!! 😄';
@@ -25,21 +42,46 @@ export class AppService {
   async createInvoice(amount: number): Promise<InvoiceResponse> {
     const invoice = this.makeInvoice(amount);
     console.log('creating invoice', invoice);
-    //   {
-    //     "id": "ecdf1c16f8ce8fe643fc9c6c76bfbef61d86a5afe5840ffbbc6cc939cbf7e542",
-    //     "status": "Unpaid",
-    //     "BOLT11": "lnbc8880n1pnk2n65pp5an03c9hce687vslun3k8d0a77cwcdfd0ukzql7audnynnjlhu4pqdqvdpshx6rtv4uscqzzsxqzz6sp509le7thnljsh0mkay4q3l6hxentwzs2v27nkpfscmmpqyxw7nf8s9qyyssqkx09jvv04a9jzwhyawt8xzkadh8qaj6dfk9863h4c3yx79m0y8zs9f65usutwyyv95tkkyvl78jj5r35qhzn4qkpthklyje76x9w9gsphm4mvt",
-    //     "paymentHash": "ecdf1c16f8ce8fe643fc9c6c76bfbef61d86a5afe5840ffbbc6cc939cbf7e542",
-    //     "paidAt": null,
-    //     "expiresAt": 1734692782,
-    //     "amount": "888000",
-    //     "amountReceived": null
-    // }
 
-    const response: InvoiceResponse = await axios.post(
-      `${this.BTCPAY_URL}/api/v1/stores/${this.BRIDGE_CENTER_ID}/lightning/BTC/invoices`,
+    const response = await axios.post(
+      `${this.BTCPAY_URL}api/v1/stores/${this.BRIDGE_CENTER_ID}/lightning/BTC/invoices`,
       invoice,
+      {
+        headers: this.getAuthHeaders(),
+      },
     );
-    return response;
+    return response.data;
+  }
+
+  async webhook(body: any): Promise<string> {
+    console.log('received webhook', body);
+    const webhookType = body.type;
+    if (webhookType === WebhookType.InvoiceCreated) {
+      // {
+      //   deliveryId: '5dGs78VTu8nmrj6rDKZXxf',
+      //   webhookId: '6mGzWTC2pYw6kg1W9GiKEe',
+      //   originalDeliveryId: '5dGs78VTu8nmrj6rDKZXxf',
+      //   isRedelivery: false,
+      //   type: 'InvoiceCreated',
+      //   timestamp: 1734685708,
+      //   storeId: 'DWZm5wfPFY2JCpf2fK6e8P7z7cr7mV1V6DQEGvC9r9EH',
+      //   invoiceId: 'YYpysWafJPwDWQtwoDBwav',
+      //   metadata: {}
+      // }
+      console.log('InvoiceCreated');
+    } else if (webhookType === WebhookType.InvoiceExpired) {
+      console.log('InvoiceExpired');
+    } else if (webhookType === WebhookType.InvoiceReceivedPayment) {
+      console.log('InvoiceReceivedPayment');
+    } else if (webhookType === WebhookType.InvoicePaymentSettled) {
+      console.log('InvoicePaymentSettled');
+    } else if (webhookType === WebhookType.InvoiceProcessing) {
+      console.log('InvoiceProcessing');
+    } else if (webhookType === WebhookType.InvoiceInvalid) {
+      console.log('InvoiceInvalid');
+    } else if (webhookType === WebhookType.InvoiceSettled) {
+      console.log('InvoiceSettled');
+    }
+    return 'test';
   }
 }
